Submap data fetching solution.
For fetching submap path, width and height .ts file was created that contains an array of objects with necessary information for each submap image. A certain submap object is being fetched through getSubmapImage function located in SubmapProviderImpl class, which implements SubmapProvider interface. The return data is represented by  newly created type Submap. One questionable aspect of this solution is the submapId variable. For better description it should probably be changed to some kind of string (open for discussion).