home page swap button, go and share buttons fixed
if dropdown with only one item was visible, click on outside would automatically fill the input. But when the "click outside" is click on some button, the button event will trigger before the autofill and it will look like a bug. Fix was to disable all three buttons while dropdown is visible